+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # [1.1.0](https://github.com/phalcongelist/breadcrumbs/releases/tag/v1.1.0) (2016-XX-XX)
 * Added support of events
+* Added `Breadcrumbs::update` to update an existing crumb
+* Added the `breadcrumbs:beforeUpdate` and `breadcrumbs:afterUpdate` events
 
 # [1.0.0](https://github.com/phalcongelist/breadcrumbs/releases/tag/v1.0.0) (2016-02-21)
 * Initial release